rename movable to slidable
